# pecos-backendadmin-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pecos River Traders API",
        "description": "API for Pecos River Traders e-commerce platform",
        "contact": {
            "email": "admin@pecosrivertraders.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/pecos-backendadmin-api/public/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/blog": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Get all published blog posts",
                "description": "Get all published blog posts.",
                "operationId": "e15c621f9e9845cfec607f9ddcdc24f7",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 9
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/blog/{slug}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Get blog post by slug",
                "description": "Get a single blog post by slug.",
                "operationId": "9299865efac92f367877874f307f239b",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            }
        },
...
        "/blog/recent": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Get recent blog posts",
                "description": "Get recent posts.",
                "operationId": "f903c8cf230388d1679b3f788bf82a79",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/admin/blog": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all blog posts (admin)",
                "description": "Admin: Get all posts (including drafts).",
                "operationId": "045328bbabe88c4fec4ba024d4edb6cc",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create blog post (admin)",
                "description": "Admin: Create blog post.",
                "operationId": "f5d3fea5208ebdbe4369f3176e1e1833",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "slug",
                                    "content",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "excerpt": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "published",
                                            "scheduled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/blog/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update blog post (admin)",
                "description": "Admin: Update blog post.",
                "operationId": "4bcc001693cfc6b74e31056bfa0517fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post updated"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete blog post (admin)",
                "description": "Admin: Delete blog post.",
                "operationId": "c60ab8cc74d781c546d5f3e950170223",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deleted"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}